% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{add_end_year_flag}
\alias{add_end_year_flag}
\title{Add the end of year flag at the end of each year in the scheduled mgt table}
\usage{
add_end_year_flag(schedule_tbl)
}
\arguments{
\item{schedule_tbl}{Tibble providing the scheduled operations for HRU i}
}
\description{
Add the end of year flag at the end of each year in the scheduled mgt table
}
\keyword{internal}
