% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgt_handling.R
\name{add_management}
\alias{add_management}
\title{Add a farm management schedule to a farmr object}
\usage{
add_management(mgt_name, mgt_schedule, cnop_table = NULL)
}
\arguments{
\item{mgt_name}{Name of the management to address in the farmr object.}

\item{mgt_schedule}{Path to a mgt schedule file in .csv format, or data.frame
(tibble) object in the working environment.}

\item{cnop_table}{Optional table with Curve number values for the operations
plant, harvest, and tillage for the respective crops in the mgt_schedule
and the present hydrologic soil groups, in .csv or data.frame format.}
}
\value{
Generates a new farmr_project in the working environment (as an R6
  object) and saves the project the TxtInOut folder.
}
\description{
Add a farm management schedule to a farmr object
}
