% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{add_management_schedule}
\alias{add_management_schedule}
\title{Add the scheduled management operations in thve mgt file}
\usage{
add_management_schedule(mgt_file, schedule_tbl)
}
\arguments{
\item{mgt_file}{Lines of the respective mgt file}

\item{schedule_tbl}{Tibble with the scheduled operations for the respective HRU}
}
\description{
Add the scheduled management operations in thve mgt file
}
\keyword{internal}
