% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{add_skip_row}
\alias{add_skip_row}
\title{Add a skip row if last row is not a skip operation}
\usage{
add_skip_row(tbl)
}
\arguments{
\item{tbl}{management input table}
}
\value{
The management input table with an added skip row
}
\description{
Add a skip row if last row is not a skip operation
}
\keyword{internal}
