% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{add_skip_year_flag}
\alias{add_skip_year_flag}
\title{Add the skip year flag for years in the scheduled mgt table where no
operations were set.}
\usage{
add_skip_year_flag(schedule_tbl, variable)
}
\arguments{
\item{schedule_tbl}{Tibble providing the scheduled operations for HRU i}

\item{variable}{Tibble of a variable required to extract all years}
}
\description{
Add the skip year flag for years in the scheduled mgt table where no
operations were set.
}
\keyword{internal}
