% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{as_num}
\alias{as_num}
\title{Convert string to numeric without warning}
\usage{
as_num(x)
}
\arguments{
\item{x}{Text string}
}
\description{
Convert string to numeric without warning
}
\keyword{internal}
