% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{assign_subbasin_weather}
\alias{assign_subbasin_weather}
\title{Assign the correct weather stations to the subbasins and return updated variables}
\usage{
assign_subbasin_weather(project_path, variables)
}
\arguments{
\item{project_path}{Path string to SWAT project folder.}

\item{variables}{List of tibbles with read weather station data}
}
\description{
Assign the correct weather stations to the subbasins and return updated variables
}
\keyword{internal}
