% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{assign_var_i}
\alias{assign_var_i}
\title{Assign the correct weather stations to the subbasins for the variable i}
\usage{
assign_var_i(var, sub, var_i)
}
\arguments{
\item{var}{List of weather variables tables.}

\item{sub}{Table that links weather stations to subbasins}

\item{var_i}{String label for the variable i}
}
\description{
Assign the correct weather stations to the subbasins for the variable i
}
\keyword{internal}
