% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{build_ini_line}
\alias{build_ini_line}
\title{Build the lines for the plant.ini text file from the prepared plant inis}
\usage{
build_ini_line(ini_i, name_i)
}
\arguments{
\item{ini_i}{Tibble with the ith plant_ini.}

\item{name_i}{Name of the corresponding operation schedule.}
}
\description{
Build the lines for the plant.ini text file from the prepared plant inis
}
\keyword{internal}
