% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{c_idx}
\alias{c_idx}
\title{Concatinate ranges of idx and idx to a vector of idx}
\usage{
c_idx(x)
}
\arguments{
\item{x}{Vector providing idx values and ranges}
}
\description{
Concatinate ranges of idx and idx to a vector of idx
}
\keyword{internal}
