% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{check_if_daily}
\alias{check_if_daily}
\title{Check if all weather inputs are in a daily time interval}
\usage{
check_if_daily(files, project_path)
}
\arguments{
\item{files}{String vector of weather input files}

\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Check if all weather inputs are in a daily time interval
}
\keyword{internal}
