% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{check_mgt_table_2012}
\alias{check_mgt_table_2012}
\title{Check the read SWAT2012 management schedule tables}
\usage{
check_mgt_table_2012(mgt_tbl, lookup, hru_attribute)
}
\arguments{
\item{mgt_tbl}{The read management table in tibble format}

\item{lookup}{List with lookup tables for plant, tillage, and fertilizer}

\item{hru_attribute}{Tibble providing attributes for the HRUs}
}
\description{
Check the read SWAT2012 management schedule tables
}
\keyword{internal}
