% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{check_var_times}
\alias{check_var_times}
\title{Check if all variables in the farmR project cover the entire time window
which was defined by the start and end years.}
\usage{
check_var_times(var_time, start_year, end_year)
}
\arguments{
\item{var_time}{Overview table of time windows for all variables}

\item{start_year}{User defined start year for operation scheduling}

\item{end_year}{User defined end year for operation scheduling}
}
\description{
Check if all variables in the farmR project cover the entire time window
which was defined by the start and end years.
}
\keyword{internal}
