% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{check_version}
\alias{check_version}
\title{Identify the SWAT version of the project path}
\usage{
check_version(project_path)
}
\arguments{
\item{project_path}{Text string path to the project TxtInOut folder}
}
\description{
Identify the SWAT version of the project path
}
\keyword{internal}
