% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{compare_reset_mgt}
\alias{compare_reset_mgt}
\title{Compare the mgt schedule which is already stored in the farmr object with
the new read schedule}
\usage{
compare_reset_mgt(
  mgt_new,
  mgt_old,
  schdl_ops,
  project_path,
  project_name,
  project_type
)
}
\arguments{
\item{mgt_new}{Tibble with new read management schedule}

\item{mgt_old}{Management schedule which was used to generate existing
scheduled operations.}

\item{schdl_ops}{List with data of previously scheduled operations}
}
\description{
Compare the mgt schedule which is already stored in the farmr object with
the new read schedule
}
\keyword{internal}
