% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{connect_unit_weather}
\alias{connect_unit_weather}
\title{Connect the HRUs to the read weather stations}
\usage{
connect_unit_weather(project_path, hru_attributes, variables, version)
}
\arguments{
\item{project_path}{Path string to SWAT project folder.}

\item{hru_attributes}{tibble for attributes of the HRUs}

\item{variables}{List of tibbles with read weather station data}
}
\description{
Connect the HRUs to the read weather stations
}
\keyword{internal}
