% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{connect_weather_2012}
\alias{connect_weather_2012}
\title{Connect the HRUs to the read weather stations for SWAT2012}
\usage{
connect_weather_2012(project_path, variables, hru_attributes)
}
\arguments{
\item{project_path}{Path string to SWAT project folder.}

\item{variables}{List of tibbles with read weather station data}

\item{hru_attributes}{tibble for attributes of the HRUs}
}
\description{
Connect the HRUs to the read weather stations for SWAT2012
}
\keyword{internal}
