% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{connect_weather_plus}
\alias{connect_weather_plus}
\title{Connect the HRUs to the read weather stations for SWAT+}
\usage{
connect_weather_plus(project_path, hru_attributes)
}
\arguments{
\item{project_path}{Path string to SWAT project folder.}

\item{hru_attributes}{tibble for attributes of the HRUs}

\item{variables}{List of tibbles with read weather station data}
}
\description{
Connect the HRUs to the read weather stations for SWAT+
}
\keyword{internal}
