% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmR_utils.R
\name{decay_i}
\alias{decay_i}
\title{Calculate decay weights.}
\usage{
decay_i(n, l)
}
\arguments{
\item{n}{Numeric value for the number of steps before or after the reference point.}

\item{l}{Numeric value for the decay rate.}
}
\description{
Calculate decay weights.
}
\keyword{internal}
