% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{document_op_skip}
\alias{document_op_skip}
\title{Document if an operation scheduling did not result in an operation date}
\usage{
document_op_skip(op_skip, attribute_hru_i, mgt_j, prev_op, j_op)
}
\arguments{
\item{op_skip}{Tibble that documents the skipped operations}

\item{attribute_hru_i}{Tibble with one line that provides the static HRU}

\item{mgt_j}{j_th line of the mgt table that should be scheduled}

\item{prev_op}{Date of the previous opeation in ymd() format.}

\item{j_op}{index of the operation in the mgt schedule table.}
}
\description{
Document if an operation scheduling did not result in an operation date
}
\keyword{internal}
