% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmR_utils.R
\name{exp_decay}
\alias{exp_decay}
\title{Cumulative exponential for a vector x.}
\usage{
exp_decay(x, n, l)
}
\arguments{
\item{x}{Input vector fro which cummulative decay is calculated.}

\item{n}{Numeric value for the number of steps before or after the reference point.}

\item{l}{Numeric value for the decay rate.}
}
\description{
Cumulative exponential for a vector x.
}
\keyword{internal}
