% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{extract_hru_2012}
\alias{extract_hru_2012}
\title{Extract HRU attributes from the SWAT2012 .hru files}
\usage{
extract_hru_2012(str_lines)
}
\arguments{
\item{str_lines}{read lines of a .hru file}
}
\description{
Extract HRU attributes from the SWAT2012 .hru files
}
\keyword{internal}
