% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{extract_sol_2012}
\alias{extract_sol_2012}
\title{Extract HRU attributes from the SWAT2012 .sol files}
\usage{
extract_sol_2012(str_lines)
}
\arguments{
\item{str_lines}{read lines of a .sol file}
}
\description{
Extract HRU attributes from the SWAT2012 .sol files
}
\keyword{internal}
