% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{extract_sol_plus}
\alias{extract_sol_plus}
\title{Extract soil names and hydrological soil groups from soils.sol}
\usage{
extract_sol_plus(sol_file)
}
\arguments{
\item{sol_file}{Character string for path to soils.sol file}
}
\description{
Extract soil names and hydrological soil groups from soils.sol
}
\keyword{internal}
