% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmr_object.R
\docType{data}
\name{farmr_project}
\alias{farmr_project}
\title{farmr_project}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
farmr_project
}
\description{
farmr_project
}
\keyword{datasets}
