% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{filter_static_rules}
\alias{filter_static_rules}
\title{Filter the operations based on the defined static rules}
\usage{
filter_static_rules(mgt_tbl, attribute_hru_i)
}
\arguments{
\item{mgt_tbl}{Tibble that provides mgt parameters and rules}

\item{attribute_hru_i}{Tibble with one line that provides the static HRU
attributes}
}
\description{
Filter the operations based on the defined static rules
}
