% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{get_station_names}
\alias{get_station_names}
\title{Retrieve the weather station names from SWAT2012 weather input file}
\usage{
get_station_names(file_path)
}
\arguments{
\item{file_path}{String path to weather file}
}
\description{
Retrieve the weather station names from SWAT2012 weather input file
}
\keyword{internal}
