% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{get_value}
\alias{get_value}
\title{Get numeric value from line in input file}
\usage{
get_value(x)
}
\arguments{
\item{x}{The text string line.}
}
\description{
Get numeric value from line in input file
}
\keyword{internal}
