% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_functions.R
\name{\%.\%}
\alias{\%.\%}
\title{Paste function pasting two character strings "." as seperator}
\usage{
a \%.\% b
}
\arguments{
\item{a}{Character string}

\item{b}{Character string}
}
\value{
"a.b"
}
\description{
Paste function pasting two character strings "." as seperator
}

