% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_functions.R
\name{\%_\%}
\alias{\%_\%}
\title{Paste function pasting two character strings "_" as seperator}
\usage{
a \%_\% b
}
\arguments{
\item{a}{Character string}

\item{b}{Character string}
}
\value{
"a_b"
}
\description{
Paste function pasting two character strings "_" as seperator
}

