% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{initialize_crop}
\alias{initialize_crop}
\title{Write values for any initial crop}
\usage{
initialize_crop(mgt_file, schedule_init)
}
\arguments{
\item{mgt_file}{Lines of the respective mgt file}

\item{schedule_init}{tibble with crop_init parameters for the respective HRU}
}
\description{
Write values for any initial crop
}
\keyword{internal}
