% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{is_identical}
\alias{is_identical}
\title{Check if two tables are identical}
\usage{
is_identical(x, y)
}
\arguments{
\item{x}{First table}

\item{y}{Second table}
}
\description{
Check if two tables are identical
}
\keyword{internal}
