% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{is_one_of}
\alias{is_one_of}
\title{Identify if x is one of the inputs provided with of_what. To be used in
pipe workflow}
\usage{
is_one_of(x, of_what)
}
\arguments{
\item{x}{Vector to be analyzed}

\item{of_what}{Vector to be compared with}
}
\description{
Identify if x is one of the inputs provided with of_what. To be used in
pipe workflow
}
\keyword{internal}
