% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{jdnyr_to_date}
\alias{jdnyr_to_date}
\title{Convert jdn and year columns to a date column}
\usage{
jdnyr_to_date(tbl)
}
\arguments{
\item{tbl}{Tibble containing the year and the jdn columns}
}
\description{
Convert jdn and year columns to a date column
}
\keyword{internal}
