% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{jdnyr_to_ymdjdn}
\alias{jdnyr_to_ymdjdn}
\title{Convert jdn and year column to year, month, day, and jdn columns}
\usage{
jdnyr_to_ymdjdn(tbl)
}
\arguments{
\item{tbl}{Tibble containing the year and the jdn columns}
}
\description{
Convert jdn and year column to year, month, day, and jdn columns
}
\keyword{internal}
