% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_simperiod.R
\name{limit_simperiod}
\alias{limit_simperiod}
\title{Limit the simulation period and write management files for this period}
\usage{
limit_simperiod(txtIO_pth, start_year, end_year, nyskip)
}
\arguments{
\item{nyskip}{}
}
\description{
Limit the simulation period and write management files for this period
}

