% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_handling.R
\name{load_demo}
\alias{load_demo}
\title{Loading SWAT demo data for testing \code{SWATfarmR}}
\usage{
load_demo(dataset, path = NULL, version = NULL)
}
\arguments{
\item{dataset}{Character string to define data set to load. Valid inputs are
\code{dataset = c('project', 'schedule')}.
\code{dataset = 'project'} loads a SWAT demo project in the defined
\code{path} location and returns the final project path as a text string in
R. The definition of the \code{version} is required.
\code{dataset = 'schedule'} writes a demo management schedule .csv file for the
demo model setup in the \code{path} location and returns the file path as a text string.}

\item{path}{Character string that defines the path where to copy the SWAT demo
project or the demo managment schedule.}

\item{version}{Character string that defines the SWAT version. Options
are \code{version = c('2012', 'plus')}. This argument is required to
load SWAT projects and shape files.}
}
\description{
With this function you can load data sets for SWAT2012 and SWAT+ demo projects.
The provided data includes demo projects, and management schedule .csv files.
}
\section{Examples}{

To learn how to load the different demo data sets with \code{SWATfarmR} see the
section \href{https://chrisschuerz.github.io/SWATfarmR/articles/SWATfarmR.html#loading-demos}{Loading demos}
on the package's \emph{Get Started} page.
}

