% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmr_handling.R
\name{load_farmr}
\alias{load_farmr}
\title{Load an existing swatfarmr project}
\usage{
load_farmr(file)
}
\arguments{
\item{file}{File path of a swatfarmr project located in a TxtInOut folder.}
}
\value{
Generates a new farmr_project in the working environment (as an R6
  object) and saves the project the TxtInOut folder.
}
\description{
Load an existing swatfarmr project
}
