% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{load_scheduled_ops}
\alias{load_scheduled_ops}
\title{Load the scheduled operations from the SQLite data base}
\usage{
load_scheduled_ops(project_path, project_name)
}
\arguments{
\item{project_path}{Path to the TxtInOut folder}

\item{project_name}{Name of the farmR project}
}
\description{
Load the scheduled operations from the SQLite data base
}
\keyword{internal}
