% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{prepare_hru}
\alias{prepare_hru}
\title{Prepare the hru-data for writing in SWAT+ and convert it to text lines}
\usage{
prepare_hru(mgt_raw, assigned_hrus)
}
\arguments{
\item{mgt_raw}{List of original files that are relevant for the mgt scheduling.}

\item{assigned_hrus}{Tibble that links the mgt schedules to the HRUs.}
}
\description{
Prepare the hru-data for writing in SWAT+ and convert it to text lines
}
\keyword{internal}
