% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{prepare_ini}
\alias{prepare_ini}
\title{Prepare the plant.ini for writing in SWAT+}
\usage{
prepare_ini(mgt_raw, schedule, start_year, end_year)
}
\arguments{
\item{mgt_raw}{List of original files that are relevant for the mgt scheduling.}

\item{schedule}{List of scheduled operations.}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Prepare the plant.ini for writing in SWAT+
}
\keyword{internal}
