% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputdata.R
\name{prepare_inputdata}
\alias{prepare_inputdata}
\title{Prepare the input data.}
\usage{
prepare_inputdata(txtIO_pth, mgt_pth, ant_days = 5)
}
\arguments{
\item{txtIO_pth}{Path for the txtInOut folder of the SWAT2012 project}

\item{mgt_pth}{Path for the folder holding the user management schedule and
CNOP .csv files}

\item{antecedent_days}{Antecendent days considered for the development of the
API (Antecedent Precipitation Index)}
}
\value{
Returns a list structure holding following components: \itemize{
  \item \code{$MGT_CNOP}: A list of the user management schedules and
     respective CNOP tables
  \item \code{$Lookup}: A lookup table holding codes for management operations,
     fertilizers, tillage operations, crop types from SWAT, Linking table for
     weather stations to Subbasins, and simulation period from file.cio.
  \item \code{$Precipitation}: Precipitation data from the SWAT project
  \item\code{$Antecedent_Precip}: Calculated antecedent precipitation, derived
     from precipitation and temperature data of the SWAT project
  \item \code{$Temperature_Ind}: Normalized temperature index calculated from
     tempearture data of the SWAT project and statistics derived from them
  \item\code{$Temperature_Stat}: Long term monthly temperature statistics,
     derived from the SWAT project temperature data.}
}
\description{
Prepare the input data.
}

