% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{prepare_lum}
\alias{prepare_lum}
\title{Prepare the landuse.lum for writing in SWAT+}
\usage{
prepare_lum(mgt_raw, schedule, assigned_hrus)
}
\arguments{
\item{mgt_raw}{List of original files that are relevant for the mgt scheduling.}

\item{schedule}{List of scheduled operations.}

\item{assigned_hrus}{Tibble that links the mgt schedules to the HRUs.}
}
\description{
Prepare the landuse.lum for writing in SWAT+
}
\keyword{internal}
