% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{prepare_mgt_i}
\alias{prepare_mgt_i}
\title{Prepare the landuse.lum for the the generated operation schedule i}
\usage{
prepare_mgt_i(schedule_i, name_i, lum_raw, start_year, end_year)
}
\arguments{
\item{schedule_i}{Tibble with the ith scheduled operations.}

\item{name_i}{Name of the ith operation schedule}

\item{lum_raw}{Initial landuse.lum}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Prepare the landuse.lum for the the generated operation schedule i
}
\keyword{internal}
