% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{prepare_plant_ini_i}
\alias{prepare_plant_ini_i}
\title{Prepare the plant.ini for the the generated operation schedule i}
\usage{
prepare_plant_ini_i(schedule_i, start_year, end_year)
}
\arguments{
\item{schedule_i}{Tibble with the ith scheduled operations.}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Prepare the plant.ini for the the generated operation schedule i
}
\keyword{internal}
