% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{prepare_variables}
\alias{prepare_variables}
\title{Prepare the variable table for the respective HRU}
\usage{
prepare_variables(var_list, hru_var_con, i_hru, swat_version)
}
\arguments{
\item{var_list}{List of tibbles that provides time series of variables for
all subbasins that can be implemented in dynamic rules.}

\item{subbasin}{Number of the subbasin in which the HRU is located}
}
\description{
Prepare the variable table for the respective HRU
}
\keyword{internal}
