% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{print_na_summary}
\alias{print_na_summary}
\title{Print a summary of NA values of weather station records}
\usage{
print_na_summary(tbl, label)
}
\arguments{
\item{tbl}{Table with weather station time series}

\item{label}{Text string to be printed in warning message}
}
\description{
Print a summary of NA values of weather station records
}
\keyword{internal}
