% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_attributes_2012}
\alias{read_attributes_2012}
\title{Read HRU attributes from the .hru and .sol files in SWAT2012 projects}
\usage{
read_attributes_2012(project_path, t0)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read HRU attributes from the .hru and .sol files in SWAT2012 projects
}
\keyword{internal}
