% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_attributes_plus}
\alias{read_attributes_plus}
\title{Extract HRU attributes for SWAT+ projects}
\usage{
read_attributes_plus(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Extract HRU attributes for SWAT+ projects
}
\keyword{internal}
