% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{read_con_file}
\alias{read_con_file}
\title{Read a SWAT+ connect file}
\usage{
read_con_file(con_path)
}
\arguments{
\item{con_path}{Path string to SWAT+ connect file.}
}
\description{
Read a SWAT+ connect file
}
\keyword{internal}
