% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{read_hru_attributes}
\alias{read_hru_attributes}
\title{Read HRU attributes from the .hru and .sol files}
\usage{
read_hru_attributes(project_path, t0)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read HRU attributes from the .hru and .sol files
}
