% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_line_file}
\alias{read_line_file}
\title{Routine to read SWAT+ input files line wise when necessary due to its structure}
\usage{
read_line_file(file_path)
}
\arguments{
\item{file_path}{String. Path to the file}
}
\description{
Routine to read SWAT+ input files line wise when necessary due to its structure
}
\keyword{internal}
