% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{read_lookup}
\alias{read_lookup}
\title{Read the lookup tables for plant, fertilizer, and tillage codes from the SWAT
project}
\usage{
read_lookup(project_path)
}
\arguments{
\item{project_path}{Text string path SWAT TxtInOut folder}
}
\description{
Read the lookup tables for plant, fertilizer, and tillage codes from the SWAT
project
}
