% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{read_mgt}
\alias{read_mgt}
\title{Read SWAT mgt input files}
\usage{
read_mgt(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read SWAT mgt input files
}
\keyword{internal}
