% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgt.R
\name{read_mgt_2012}
\alias{read_mgt_2012}
\title{Read SWAT mgt input files for SWAT2012}
\usage{
read_mgt_2012(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read SWAT mgt input files for SWAT2012
}
\keyword{internal}
