% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{read_mgt_lkp}
\alias{read_mgt_lkp}
\title{Read the management schedule table from a csv file}
\usage{
read_mgt_lkp(file, version, project_path, hru_attribute)
}
\arguments{
\item{file}{Text string path to the csv file}

\item{version}{String that indicates what SWAT version the project is.}
}
\description{
Read the management schedule table from a csv file
}
\keyword{internal}
