% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgt.R
\name{read_mgt_plus}
\alias{read_mgt_plus}
\title{Read SWAT mgt input files for SWAT+}
\usage{
read_mgt_plus(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read SWAT mgt input files for SWAT+
}
\keyword{internal}
