% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{read_mgt_table}
\alias{read_mgt_table}
\title{Read the management schedule table from a csv file}
\usage{
read_mgt_table(file)
}
\arguments{
\item{file}{Text string path to the csv file}
}
\description{
Read the management schedule table from a csv file
}
