% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{read_mgt_table_2012}
\alias{read_mgt_table_2012}
\title{Read the management schedule table from a csv file for SWAT2012}
\usage{
read_mgt_table_2012(file)
}
\arguments{
\item{file}{Text string path to the csv file}
}
\description{
Read the management schedule table from a csv file for SWAT2012
}
\keyword{internal}
