% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{read_plt_comm}
\alias{read_plt_comm}
\title{Read the SWAT+ plant.ini file}
\usage{
read_plt_comm(comm_file)
}
\arguments{
\item{comm_file}{Text string path to the plant.ini file}
}
\description{
Read the SWAT+ plant.ini file
}
\keyword{internal}
