% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_rtu_def}
\alias{read_rtu_def}
\title{Routine to read the SWAT+ routing unit definition}
\usage{
read_rtu_def(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Routine to read the SWAT+ routing unit definition
}
\keyword{internal}
