% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_table_linewise}
\alias{read_table_linewise}
\title{Read text file tables safely linewise and then convert to table}
\usage{
read_table_linewise(file_path, col_names, col_types, n_skip)
}
\arguments{
\item{file_path}{String. Path to the file}

\item{col_names}{Vector with column names. If 'get' the names are
estracted from tne n_skip - 1 line.}

\item{col_types}{Vector with col types. 'c' for character, 'd' for double}

\item{n_skip}{Number of lines to skip}
}
\description{
Read text file tables safely linewise and then convert to table
}
\keyword{internal}
