% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{read_weather}
\alias{read_weather}
\title{Read a weather input file}
\usage{
read_weather(file, var, skip, digit_var, digit_date)
}
\arguments{
\item{file}{Path string to weather input file.}

\item{var}{Character vector that defines the variables}

\item{skip}{Integer to skip the header}

\item{digit_var}{Integer, number of digits per variable}

\item{digit_date}{Vector of length 2 Integers give digits of year and jdn}
}
\value{
Generates a new farmr_project in the working environment (as an R6
  object) and saves the project the TxtInOut folder.
}
\description{
Read a weather input file
}
