% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{read_weather_2012}
\alias{read_weather_2012}
\title{Read the weather data for a SWAT2012 project}
\usage{
read_weather_2012(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read the weather data for a SWAT2012 project
}
\keyword{internal}
