% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{read_weather_plus}
\alias{read_weather_plus}
\title{Read the weather data for a SWAT+ project}
\usage{
read_weather_plus(project_path)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}
}
\description{
Read the weather data for a SWAT+ project
}
\keyword{internal}
