% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{reset_lum_plus}
\alias{reset_lum_plus}
\title{Reset the file 'landuse.lum' file of a SWAT+ setup to the initial file}
\usage{
reset_lum_plus(mgt_raw)
}
\arguments{
\item{mgt_raw}{List of original mgt files}
}
\description{
Reset the file 'landuse.lum' file of a SWAT+ setup to the initial file
}
\keyword{internal}
