% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{reset_mgt}
\alias{reset_mgt}
\title{Reset all management files back to the original files}
\usage{
reset_mgt(path, mgt_raw)
}
\arguments{
\item{path}{Tibble Path to the TxtInOut folder}

\item{mgt_raw}{List of original mgt files}
}
\description{
Reset all management files back to the original files
}
\keyword{internal}
