% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{sample_management}
\alias{sample_management}
\title{Sample management type if different management types are provided for the
same land use}
\usage{
sample_management(mgt_tbl)
}
\arguments{
\item{mgt_tbl}{Tibble that provides mgt parameters and rules}
}
\description{
Sample management type if different management types are provided for the
same land use
}
