% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{schedule_date_j}
\alias{schedule_date_j}
\title{Schedule the operation date for the operation j}
\usage{
schedule_date_j(var_tbl, eval_str, prev_op)
}
\arguments{
\item{var_tbl}{Tibble with the variable time series for the HRU}

\item{eval_str}{Text string expression of the dynamic rule that is evaulated
to select the date.}

\item{prev_op}{Date of the previous opeation in ymd() format.}
}
\description{
Schedule the operation date for the operation j
}
