% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{schedule_init}
\alias{schedule_init}
\title{Schedule the initial crop operation}
\usage{
schedule_init(mgt_op, version)
}
\arguments{
\item{mgt_op}{Tibble with one line for the current operation}
}
\description{
Schedule the initial crop operation
}
\keyword{internal}
