% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{schedule_op_j}
\alias{schedule_op_j}
\title{Add the operation j in the opeation schedule for HRU i}
\usage{
schedule_op_j(schedule_i, mgt_j, date_j)
}
\arguments{
\item{schedule_i}{Scheduled operations table for the HRU i}

\item{mgt_j}{j_th line of the mgt table that should be scheduled}

\item{date_j}{The calculated date for the mgt operation j}
}
\description{
Add the operation j in the opeation schedule for HRU i
}
