% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{schedule_operation}
\alias{schedule_operation}
\title{Schedule management operations for all HRUs based on static and dynamic rules}
\usage{
schedule_operation(mgt_schedule, hru_attribute, variables, lookup)
}
\arguments{
\item{mgt_schedule}{Tibble that provides mgt parameters and rules}

\item{hru_attribute}{Tibble that provides static HRU attributes}

\item{variables}{List of tibbles that provides time series of variables for
all subbasins that can be implemented in dynamic rules.}

\item{lookup}{list of tibbles with lookup tables for plant, tillage,
fertilizer codes from SWAT data base.}
}
\description{
Schedule management operations for all HRUs based on static and dynamic rules
}
