% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{set_colnames}
\alias{set_colnames}
\title{Set colnames of a data frame or matrix}
\usage{
set_colnames(x, nm)
}
\arguments{
\item{x}{Data frame or matrix}

\item{nm}{Character names vector}
}
\description{
Set colnames of a data frame or matrix
}
\keyword{internal}
