% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{set_skip_flag}
\alias{set_skip_flag}
\title{Set the skip flag for entries in the list of tibbles with only one row and
all NA values for operation}
\usage{
set_skip_flag(tbl)
}
\arguments{
\item{tbl}{Tibble in the list of tibbles grouped by years}
}
\description{
Set the skip flag for entries in the list of tibbles with only one row and
all NA values for operation
}
