% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_check_management.R
\name{translate_mgt_table_2012}
\alias{translate_mgt_table_2012}
\title{Translate the operation and management labels into SWAT codes}
\usage{
translate_mgt_table_2012(mgt_tbl, lookup)
}
\arguments{
\item{mgt_tbl}{Loaded tibble with the management operation schedules}

\item{lookup}{List of lookup tables}
}
\description{
Translate the operation and management labels into SWAT codes
}
\keyword{internal}
