% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmR_utils.R
\name{w_log}
\alias{w_log}
\title{Logistic function to calculate weight transition between boundaries.}
\usage{
w_log(x, lwr, upr)
}
\arguments{
\item{x}{Numeric value to calculate weight from}

\item{lwr}{Numeric value for lower boundary. If \code{x = lwr} the function returns ~0.01 as a weight.}

\item{upr}{Numeric value for upper boundary. If \code{x = upr} the function returns ~0.99 as a weight.}
}
\description{
Logistic function to calculate weight transition between boundaries.
}
\examples{
\dontrun{
# All examples below must be included in the dynamic
# rules in the management csv file to be effective.

# Low weight for days below 0 degC with a
# logistic transition from 5 degC to 0 degC:
tmp <- -5:10
weight <- logistic(tmp, 0, 5)
plot(tmp, weight)

# Low weight for days with high precipitation
# (transition range: 2 to 10 mm)
pcp <- 0:12
weight <- (1 - logistic(pcp, 0, 5))
plot(pcp, weight)

# Harvesting crop in optimum PHU fraction range 1 to 1.1
hu_fr <- seq(0.75, 1.4, 0.01)
weight <- logistic(hu_fr, 0.9, 1) * (1 - logistic(hu_fr, 1.1, 1.2))
}

}
