% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{write_file_cio}
\alias{write_file_cio}
\title{Write the start and end dates into the file.cio}
\usage{
write_file_cio(path, start_year, end_year)
}
\arguments{
\item{path}{Path to the TxtInOut folder}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Write the start and end dates into the file.cio
}
\keyword{internal}
