% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_luctemp.R
\name{write_luctemp}
\alias{write_luctemp}
\title{Write a template for the luc table of update_landuse()}
\usage{
write_luctemp(luc_path = NULL)
}
\arguments{
\item{luc_path}{Either path string to write a .csv or NULL. Then the
template structure is returned as a data.frame}
}
\description{
Write a template for the luc table of update_landuse()
}
\examples{
# To get a template for the luc_tbl

write_luctemp(choose.dir()) # only windows
write_luctemp() #Returns tibble to .Globalenv
}

