% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_luctemp.R
\name{write_luptemp}
\alias{write_luptemp}
\title{Write a template for the lup table of update_landuse()}
\usage{
write_luptemp(lup_path = NULL)
}
\arguments{
\item{lup_path}{Either path string to write a .csv or NULL. Then the
template structure is returned as a data.frame}
}
\description{
Write a template for the lup table of update_landuse()
}
\examples{
# To get a template for the lup_tbl

write_luptemp(choose.dir()) # only windows
write_luptemp() #Returns tibble to .Globalenv
}

