% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgtbackup.R
\name{write_mgtbackup}
\alias{write_mgtbackup}
\title{Write management backup}
\usage{
write_mgtbackup(txtIO_pth, overwrite = FALSE)
}
\arguments{
\item{txtIO_pth}{Path to SWAT project}

\item{overwrite}{Logical, TRUE for overwrite of old managament files}
}
\description{
Write management backup
}

