% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{write_op_2012}
\alias{write_op_2012}
\title{Write the management schedules for a SWAT2012 project}
\usage{
write_op_2012(
  path,
  proj_name,
  mgt_raw,
  assigned_hrus,
  schedules,
  start_year,
  end_year
)
}
\arguments{
\item{path}{Path to the TxtInOut folder}

\item{mgt_raw}{List of original files that are relevant for the mgt scheduling.}

\item{assigned_hrus}{Tibble that links the mgt schedules to the HRUs}

\item{schedules}{List of tibbles with the scheduled operations.}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Write the management schedules for a SWAT2012 project
}
\keyword{internal}
