% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{add_variable}
\alias{add_variable}
\title{Add additional variables that can be used to define rules for}
\usage{
add_variable(data, name, assign_unit, con, variables)
}
\arguments{
\item{data}{The table that should be added as a variable. The number of columns
must be 1 or the number of subbasins. the row number must be the same as the
number of weather records.}

\item{name}{Character string to define the name of the added variable (this
name must be used in the rule set)}

\item{assign_unit}{tibble to link spatial units to the variables in data}

\item{variables}{Internal variables. Not defined by user. Required number of variables.}

\item{date}{Internal variable. Not defined by user. Date vector that is added.}
}
\value{
Generates a new farmr_project in the working environment (as an R6
object) and saves the project the TxtInOut folder.
}
\description{
Add additional variables that can be used to define rules for
}
\keyword{internal}
