% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{extract_hru_attr}
\alias{extract_hru_attr}
\title{Extract HRU attributes from the .hru files}
\usage{
extract_hru_attr(str_lines)
}
\arguments{
\item{str_lines}{read lines of a .hru file}
}
\description{
Extract HRU attributes from the .hru files
}
