% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat_input.R
\name{extract_sol_attr}
\alias{extract_sol_attr}
\title{Extract HRU attributes from the .sol files}
\usage{
extract_sol_attr(str_lines)
}
\arguments{
\item{str_lines}{read lines of a .sol file}
}
\description{
Extract HRU attributes from the .sol files
}
\keyword{internal}
