% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmr_object.R
\name{farmr_project}
\alias{farmr_project}
\title{farmr_project}
\description{
farmr_project

farmr_project
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-farmr_project-new}{\code{farmr_project$new()}}
\item \href{#method-farmr_project-add_variable}{\code{farmr_project$add_variable()}}
\item \href{#method-farmr_project-read_management}{\code{farmr_project$read_management()}}
\item \href{#method-farmr_project-save}{\code{farmr_project$save()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farmr_project-new"></a>}}
\if{latex}{\out{\hypertarget{method-farmr_project-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farmr_project$new(project_name, project_path, project_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farmr_project-add_variable"></a>}}
\if{latex}{\out{\hypertarget{method-farmr_project-add_variable}{}}}
\subsection{Method \code{add_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farmr_project$add_variable(data, name, assign_unit = NULL, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farmr_project-read_management"></a>}}
\if{latex}{\out{\hypertarget{method-farmr_project-read_management}{}}}
\subsection{Method \code{read_management()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farmr_project$read_management(file, discard_schedule = "no")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farmr_project-save"></a>}}
\if{latex}{\out{\hypertarget{method-farmr_project-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farmr_project$save()}\if{html}{\out{</div>}}
}

}
}
