% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_operations.R
\name{prepare_schedule_con}
\alias{prepare_schedule_con}
\title{Prepare the management schedule connection table}
\usage{
prepare_schedule_con(hru_attributes, swat_version)
}
\arguments{
\item{hru_attributes}{Tibble that provides the HRU attributes}

\item{swat_version}{label indicating the SWAT version}
}
\description{
Prepare the management schedule connection table
}
\keyword{internal}
