% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hru_attributes.R
\name{read_hru_attributes}
\alias{read_hru_attributes}
\title{Read the weather data}
\usage{
read_hru_attributes(project_path, version, t0)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}

\item{version}{String that indicates what SWAT version the project is.}

\item{t0}{date time object storing initial start of initialization}
}
\description{
Read the weather data
}
\keyword{internal}
