% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgt.R
\name{read_mgt}
\alias{read_mgt}
\title{Read the raw management data for backup}
\usage{
read_mgt(project_path, version)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}

\item{version}{String that indicates what SWAT version the project is.}
}
\description{
Read the raw management data for backup
}
\keyword{internal}
