% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variable.R
\name{read_weather}
\alias{read_weather}
\title{Read the weather data}
\usage{
read_weather(project_path, version)
}
\arguments{
\item{project_path}{String. Path to the TxtInOut folder of the SWAT project}

\item{version}{String that indicates what SWAT version the project is.}
}
\description{
Read the weather data
}
\keyword{internal}
