% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_operations.R
\name{write_operation}
\alias{write_operation}
\title{Write the mgt schedules into the respective files in the TxtInOut folder}
\usage{
write_operation(
  path,
  proj_name,
  mgt,
  mgt_raw,
  assigned_hrus,
  start_year,
  end_year,
  year_range,
  version
)
}
\arguments{
\item{path}{Path to the TxtInOut folder}

\item{mgt_raw}{List of original files that are relevant for the mgt scheduling.}

\item{assigned_hrus}{Tibble that links the mgt schedules to the HRUs}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}

\item{year_range}{numeric vector with start and end year of the scheduled operations.}

\item{version}{String that indicates the SWAT version.}

\item{schedule}{List of tibbles with the scheduled operations.}
}
\description{
Write the mgt schedules into the respective files in the TxtInOut folder
}
\keyword{internal}
